refactor(evalforge-core): move schema, evalforge client, mapper, and auth into the package
