+++ packages/evalforge-core/src/index.ts
@@ -1,1 +1,4 @@
-export const EVALFORGE_CORE_WIRED = true;
+export * from './schema';
+export * from './evalforge';
+export * from './evalforge-mapper';
+export * from './auth';
